Backport ENG-12697: One of the UpdateClasses optimizations skips HSQLDB compiler reconstruction and also reuse the AdHoc planner cache underneath.
If there is any following AdHoc queries that hit the old AdHoc Planner thread, the cached AdHoc query plan that contains the catalog hash will not match the most recent catalog hash. Thus, the AdHoc query planner thread will keep replanning the work and not return to the client.

Impact:
If this is the case, any following AdHoc queries, DDL changes, deployment changes on the same node will not be executed. Queries issued on the other node can be executed successfully.
Workaround:
Node rejoin and fix the issue, but this is not a good solution.

Fixes:

Any non-schema changes that reuse the Hsqldb compiler should clear the AdHoc queries cache.
or
Update the cached plan with the most recent catalog hash, which can still reuse the plan.
Reasons for regression:

Not fully understand the impact of reusing Hsqldb compiler.
Testing hole related to UpdateClasses.