Disable an assertion with valgrind builds.
There is an assertion that truncating a table with a view can fail when
the result of the view update may violate a constraint.  With normal
debug and release builds this is just right, since we delete row by row.
But for valgrind builds we delete all rows at once, which means the
views are not updated until the source tables are all truncated.  So the
constraint violations are not triggered.  Sicne the test was looking for
a violation the test failed under memcheck (== valgrind) builds.