ENG-11823: Fix concurrent BulkLoader accounting error.
We allow creating multiple VoltBulkLoaders in the same Java client. They will
share some resources like the PerPartitionTable instances. However, when a batch
of rows are successfully loaded, the callback function only updates the
accounting in one of the VoltBulkLoaders in that batch. This resulted in some
weird values in outstanding row count and completed row count in VoltBulkLoader,
causing VoltBulkLoader.drain() to hang.

This is fixed now by updating the accounting in each VoltBulkLoader involved in
a batch with the proper count.