ENG-12542: VoltdbInit-Relative-Path (#4583)
Init with absolute path: Relative paths from the deployment file become absolute. Default path values are absolute too.

Init with relative path: Paths from the deployment file are preserved. Default path values are set to be relative to the voltdbroot. The voltdbroot path itself is absolute.

Start with absolute path: Path properties stay the same as they are in the "path.properties" file.

Start with relative path: Path properties stay the same as they are in the "path.properties" file.

Get with absolute path: Paths in the deployment file are the same as those in the voltdbroot.

Get with relative path: Paths in the deployment file are the same as those in the voltdbroot.