ENG-11907: (#4633)
It is possible to promote the DRConsumerMpCoordinator while a restarted ApplyBinaryLogMP txn is still executing. If the promotion occurs before the restarted txn completes, the trackers used by promoted BufferReceivers will not reflect the restarted ApplyBinaryLog and DR Consumer will stop processing logs. This patch adds a loopback that does nothing to the @ExecuteTask SystemStoredProcedure. The loopback just runs on the MPI and than returns a response. If this SPI completes, it means that any restarted MP TXNs have also completed and that it is safe to inform BufferReceivers that the DRConsumerMPCoordinator is ready to resume.