ENG-12034 address a multiple node kill scenario where the Zookeeper data went stale midway through sending the importer channel assignments. (#4680)
When the leader is distributing a new set of importer channel assignments, it will retry sending ALL assignments if ANY fail. This is fine if the first assignment fails, but in this scenario there is a race condition between the distribution of channel assignments and a mesh monitor responding to the node failure.

In the original problem report from our endurance tests, the leader has sent roughly 2/3 of the assignments out before Zookeeper throws a "bad version" error. When the entire re-send gets retried, the very first assignment gets flagged as a duplicate resulting in the reported error. The retry is aborted, meaning the remaining channel assignments don't get distributed.

Anyone who sees this in production should first be sure the cluster didn't initiate a global crash (if it did, all nodes will be gone soon and there are bigger issues). Then pause the cluster, verify the pause has completed (due to an unrelated issue), and then resume the cluster. This will restart the importers, refreshing their channel assignments. The data which was unable to import will remain in Kafka - data won't be lost unless Kafka deletes it.