ENG-12070 don't throw IllegalStateException when importers are shut down (#4682)
If an importer has not fully initialized, a callback will not be registered for it. There are a few scenarios where this can happen - see ImportManager.readyForData for them. This results in a benign error being thrown on importer shutdown, such as when the cluster is paused.

In the benign scenario, log a NOTICE and do nothing. An exception will still get thrown if the importer assignments are ever non-empty.

Unit test added to catch the issue reliably, since the issue is easiest to reproduce if Kafka doesn't actually exist. System tests see this once in awhile as well. They don't explicitly try to shut down clusters with importers before data gets processed, but due to the stressful conditions of those tests it happens anyway.

Changes to improve the usability of the LocalCluster log scraper API are included in this change - be very careful if you try to revert it! I made the API a bit more flexible (this only affects tests).