ENG-12510: Implement 2 versions of ElasticChange event. One version adds elastic change events on existing SP partitions and all MP (including the newly added hosts). The other version (which is commented out) has the same behaviour but does not have the special handling for the replicated table stream. This code can be enabled when the Replicated Table Stream is removed from DR.
We now also update the total partitions in the DRProducer when the Elastic Change event is generated. All of this is done as a new step in the ElasticJoinCoordinator. This step is done after the replicated snapshot finishes but before the index generation. The ElasticJoinCoordinator is recovered or moved due to a node failure, the restart of the coordinator will check if this step has been completed or not. The completion check is done by determining if the (MP) ExecuteTask completed by checking if the current DRId in any of the new partitions is greater than -1.