Unify temp tuple handling in the scan executors.  Add unit tests.
The scan executors need a tuple to hold the output of an inline select
into insert tuple. They did exactly the same thing, so it made sense to
make it part of initializing the insert executor.  This is what the
aggregate executor does as well.

Also, I added unit tests for the original problem and for a new wrong
answer problem I discovered while wandering aimlessly in this desert.