Fix eecheck error reporting.
We used to use --error-exitcode=-1 to force valgrind to return -1 on
error.  Valgrind returned 0 in this case.  We now use
--error-exitcode=1, which seems to work better.  We now also process the
xml output files always, and complain if the return value is non-zero
but we don't find errors in the xml file, or zero and we do find errors.
We also print the valgrind command line.