Fixed memory leaks in expression_test and generated tests with strings.
The C++ test expression_test.cpp caused a memory leak.  Unary minus only
uses the left operand.  We push a dummy expression for unary minus to
stand as a right operand. We used to push a real expression, but now we
just push NULL.

The generated tests which use strings did not free memory when they
exit.

These are both just test errors, and not product errors.