Eng 12481 test query timeout (#4717)
We stand up a server with authentication enabled.  When we run the test
in debug we don't actually do any testing.  But in this case we still
need to set the authentication credentials after each test, because we
need these to be right in order to connect to the server to tear it
down.

Note that this test really does not run in debug.

https://issues.voltdb.com/browse/ENG-12481