Eng 12901 Valgrind Botch
We used to use --error-exitcode=-1 to force valgrind to return -1 on
error.  Valgrind returned 0 in this case.  We now use
--error-exitcode=1, which seems to work better.  We now also process the
xml output files always, and complain if the return value is non-zero
but we don't find errors in the xml file, or zero and we do find errors.
We also print the valgrind command line.

The C++ test expression_test.cpp caused a memory leak.  Unary minus only
uses the left operand.  We push a dummy expression for unary minus to
stand as a right operand. We used to push a real expression, but now we
just push NULL.

The generated tests which use strings did not free memory when they
exit.

Fix the CompactionTest and CopyOnWriteTest to run with valgrind.

These are all just test errors, and not product errors.

https://issues.voltdb.com/browse/ENG-12901