Add an MP procedure to calculate wait time by staion in a time window.
Call the procedure periodically to calculate the wait time for each station and
store them in a table. Query the view wait_time_by_station to get the total wait
time and total entries to calculate the average wait time.

The wait_time table is pruned when you call UpdateWaitTime. The time window is
set to 5 minutes in the past.