Eng 12528 file export delimiter patch (#4698)
Character <, >, & should be encoded as XML entities.

Control characters should be encoded as Java literals. For example, you should put \1 for Ctrl-A character and \n for new line character. If you want to use space as the last delimiter, please encode it as \u0020 ( the Unicode representation of space character).