Initialize the function table from the catalog.
We store the set of UDFs in the catalog.  When we start to compile, we
will clear the UDF table in the compiler and reinitialize it from the
catalog.  This means that the UDF table is always up to date when we
start compiling, and the function IDs will be stable, even if AdHoc
compilations are done on unpredictable nodes.