Saved before federation.
I'm going to try to make this right.  The set of function names will not
be all stored statically in FunctionForVolt.FunctionId.  There will be a
FunctionForVoltDB. FunctionId in each HsqlDbInterface object.  Most
functions will be static and shared between the instances.  But UDFs
will be stored in HSQLDBInterface object.  The namespace of SQL
functions will be federated between the static functions and the UDFs.