Eng 13032 timeout ntprocedure (#4827)
ENG-13032: There is an edge case that a MPI node may fail while running UpdateClasses and the ZK UAC blocker is not removed. This is possible if this failed MPI node has not received any invocation of @updatecore.

The fix is to get the completable future with a 10 seconds time out. If the future is not complete, we will remove the UAC blocker anyway and return a time out error message to client. The UpdateCore may eventually succeed even client receiving a timeout response. This will guarantee the removal of UAC blocker.