Fix interactive inline batch bug.
The command "file -inlinebatch EOB" failed interactively.  This is
because we create context to remember the name of the input file.  But
an interactive command execution has no input file, so this context is
null.  The NPE is not handled properly.

I don't actually know how to test this other than manually.  The problem
does not show up in sqlcmdtest, since this input is not interactive.