ENG-13039, partially revert ENG-11743 to fix a race condition that could leads to StatsAgent thread hang forever. (#4849)
* ENG-13039, partially revert ENG-11743 to fix a race condition that could leads to StatsAgent thread hang forever.

During schema change, Site thread starts to update all the related data structures, including clearing the stats sources in StatsAgent class, but there is a race window that in the meantime StatsAgent thread is iteratoring the stats sources to response a statisticss request. Deleting the procedure statisticss source while StatsAgent thread is iteratoring may fooled this thread to think now is in startup time and thus it's better to yield CPU for a little while and retry later, which in current case is hang forever.

ENG-11743 introduced this race window, it is partially reverted because we no longer need to keep @UpdateCore stats during schema change.