ENG-13045: Be more careful about incorrectly reporting success after a snapshot. Specifically, the last reporting host could overwrite a failure with success for snapshot completion listeners such as CommandLogging. In CommandLogging's case, this could cause a trucation even though the snapshot did not complete successfully.
Another case was SnapshotStatus Statistics, which could report a successful snapshot if a failure occured during file open or the writing of the initial table header. This is because a header write failure caused the NativeSnapshotTarget to be replaced with the DevNullSnapshotTarget, which did not preserve the earlier exception. Now, DevNullSnapshotTarget can optionally be constructed to include and propagate earlier exceptions to snapshot statistics so that snapshot stats correctly mark the result for that file as "FAILURE".