Changed to a federated name table for HSQLDB.
SQL functions are now federated between built-in functions and user
defined functions.  There are two tables in FunctionsForVolt, one for
each.  When compiling DDL, we save away the old user defined functions
and move them back when they are redefined.  If a procedure compilation
fails, we can just restore the saved function definitions by moving a
single map object.  This is fast and simple, at the cost of an extra
lookup for function names.