ENG-12282: Fix export promotion and ack race condition.
A race condition between export data source master promotion and getting acks
from the previous failed master can occur. The failed master could have sent out
an ack with Long.MIN and fails immediately after that, which causes a new master
to be elected. The election and the receiving of this ack message happens on two
different threads on the new master. If it's promoted while processing the ack,
the ack may call `m_onDrain` while the other thread is polling buffers, which
may never get discarded.

Check to see if we are already promoted to be the master. If so, ignore the ack.