Eng 13045 fail host on truncation snapshot failure (#4855)
* ENG-13045:
Make sure that any IO errors generated for a Truncation Snapshot brings down the host. Detect IO errors on table, digest, completion, and catalog files.

Add Unit test that uses hacks in table file construction to force an IO exception. Verify that this generates a crash for truncation snapshots only.

* ENG-13045:
Be more careful about incorrectly reporting success after a snapshot. Specifically, the last reporting host could overwrite a failure with success for snapshot completion listeners such as CommandLogging. In CommandLogging's case, this could cause a trucation even though the snapshot did not complete successfully.

Another case was SnapshotStatus Statistics, which could report a successful snapshot if a failure occured during file open or the writing of the initial table header. This is because a header write failure caused the NativeSnapshotTarget to be replaced with the DevNullSnapshotTarget, which did not preserve the earlier exception. Now, DevNullSnapshotTarget can optionally be constructed to include and propagate earlier exceptions to snapshot statistics so that snapshot stats correctly mark the result for that file as "FAILURE".