Add support for UDF dependences in procedures with DML.
I addressed some of Ethan's comments about procedures with insert and
delete statements.  I added unit tests to verify that we do the right
thing when trying to delete UDFs which procedures depend on if the
procedures have insert or delete statements.

Also, some small fixes for functions in the order by list of delete
statements.  This was broken in master.