ENG-13168: Make TrackerHelper aware of DRSiteDrIdTracker (that includes lastReceivedLogId).
TestDR2Reset relies on new CLI because old CLI has the side-effect, in LocalCluster, of deleting all subRoots (including other instances of LocalCluster). There is now an explicit check in the Test and LocalCluster has been updated to use new CLI by default if NEW_CLI is not specified as a System or JVM variable.