Eng 13168 monatomic last received (#4936)
* ENG-13168:
Instead of using the last received DRId to detect duplicate ApplyBinaryLogs in the Site, we will use a new monatomicly incremented value that starts at 0 when the receiver for a cluster is first initialized. Since we only need duplicate detection for Leader failover, this value does not need to be durable. However, because rejoined hosts use snapshots to synchronize partition state, this value is durable anyway

* ENG-13168: Initialize Site counter and Partition counters to 0, implying that the first valid logId will be 1.

* ENG-13168: All tracker's last received should be initialied to 0 because the first request (usually initialized tracker from StartStream) should use logId of 1.

* ENG-13168: Make sure the UndoActions list in Java are removed after an Undo so the actions don't get executed twice when a TXN is restarted.

* ENG-13168:
Make TrackerHelper aware of DRSiteDrIdTracker (that includes lastReceivedLogId).

TestDR2Reset relies on new CLI because old CLI has the side-effect, in LocalCluster, of deleting all subRoots (including other instances of LocalCluster). There is now an explicit check in the Test and LocalCluster has been updated to use new CLI by default if NEW_CLI is not specified as a System or JVM variable.