ENG-12827 ENG-12736 ENG-12913: DR protocol version change and various other fixes (#4858)
For ENG-12827, Added a constant NO_REPLICATED_STREAM_PROTOCOL_VERSION in DRProtocol to control if the replicated stream is created or not. Currently, it is set to 9 while the latest protocol version is 8, so replicated stream is still there now. Any if checks that use this constant if part of the work to make DR work regardless of the existence of the replicated stream. Note that the no replicated table stream change is currently disabled in this commit for stability reason.

For ENG-12736, Changed DRProducer to record and persist the negotiated version with the first consumer cluster. Any later consumer clusters with protocol version lower than this version will be denied for connecting.

For ENG-12913, create a ZK blocker for elastic join if the first negotiated version is a version that does not support elastic join during DR (currently only 7), and remove the ZK blocker when all conversations with this negotiated version are reset. Also, disallow consumers with version 7 to connect if the producer cluster has been elastically modified.