ENG-13083: Track Dependences between UDFs and Procedures. (#4879)
We now track dependences between User Defined Functions and statements in procedures.

These dependences are kept in the catalog.  For each UDF there is a list of procedure/statement
pairs depending on the UDF, and for each statement there is a list of UDFs the statement depends
on.  A UDF cannot be dropped if a statement depends on it.

https://issues.voltdb.com/browse/ENG-13083