ENG-13279: (#4967)
When the Consumer is shut down, due to a reset, a disable, or an ElasticAdd, the objects are not garbage collected because the references for each BufferReceiver are maintained in the Scheduler's DurabilityListener, even when new instances are started. This is fixed by unregistering the listener from the BufferReceiver's shutdown path. Note that the receivers and dispatcher are only leaked when commandlogging is Asynchronous and in a same-size cluster.