ENG-13201: Fix DR buffer expansion (#4946)
Now allows a single extendBufferChain for defaultSize < minLength <= maxSize.

DR should allow changes where a single tuple size exceeds the default size but is less than the max size. Before this change, extendBufferChain in TupleStreamBase checked if the requested extension is greater than the default size and failed if it was, which is incorrect behavior for DR. This PR changes TupleStreamBase to support maxSize also and changed extendBufferChain accordingly.