Eng 10092 Fix misoriented polygons from polygonfromtext. (#4911)
Fix misoriented polygon loops.

If a loop is incorrectly oriented, the function polygonfromtext() simply inserts it, resulting in invalid data, but the insertion is fast.  The function validpolygonfromtext() used to throw an exception in this case.  Now validpolygonfromtext() tries to correct these orientation errors.  If the polygon is invalid for some other reason, say edges cross, validpolygonfromtext() throws a SQL exception.  The function makevalidpolygon() takes a polygon object as a single argument and tries to correct orientation errors.  If the input argument is correctly oriented the output is a copy of the input.  If the input argument is misoriented the output argument is the result of reversing loops until the errors are corrected.

https://issues.voltdb.com/browse/ENG-10092