Don't double reverse rings in add.
The OGC WKT format requires holes to be CW, but S2 wants them to be CCW.
We reverse these for the Java representation and re-reverse them when
creating WKT.  When creating a GeographyValue from a list of lists of
points we want the input to be the WKT format.  But for the add function
we want the input to be the S2 format, since the polygon we are adding
to will be in the S2 format.  So we need to know which case we have.

https://issues.voltdb.com/browse/ENG-12926