ENG-13366: Fix XDCR conflict log in mixed version mode.
The schema of the conflict export table has changed without updating the
compatibility binary log sink code. Whenever there is a conflict using mixed
version XDCR, it fails to write the conflict and breaks replication. I brought
all the changes in the original commit that updated the schema to the
compatibility mode. The commit that introduced the change was
91a0c902d31d4830313e782403d678225a078683. See the commit message for details of
what's included.