ENG-12540: (#5034)
PlanNodes allocate TempTables during execution. This can be problematic when making Replicated Table changes because the TempTable rows are allocated in an MP memory context but are cleaned up in the Partitioned memory context during UAC or Shutdown. To resolve this we now only assume the memory context inside the executor either before or after touching the temp table.

Also fixed another unit test that manuplates multiple engines directly by assigning the engines explicitly to threads.