sql_grammar_generator.py: added a new --find_in_log option, which allows (#5053)
you to search the log (and/or console) after each SQL statement (or
other command) is issued to sqlcmd, to see which statements are causing
certain phrases of interest (e.g. NullPointerException) to appear there,
and output the result in a file(s); note that this is slow, so should
not be used routinely, only when tracking down specific, known issues;
also related options --find_files, to specify the file(s) (defaults to
existing LOG_FILES option value) and --find_number, to specify the
number of lines of the log file(s) to be checked via a 'tail' command
(defaults to 100); added a bunch of new types of ERROR to grep for in
the log (& console), including 'is already defined', 'is not defined',
and 'WARN', among others; added a small kludge to workaround ENG-13416;
modified last_n_sql_statements to be dictionary rather than a list,
which is a lot clearer; tweaked the comments for a bunch of the existing
options.

run.sh: minor tweaks to checking whether a VoltDB server is already
running (don't confuse it with SQLCommand); and to running the actual
tests (make sure the command echoed is the same one that is running,
including the path).