More refactoring to plan with CommonTable scans.
I am generating the plans now.  I still need to deal with
(1) partitioning, though this is very restricted, (2) widening
of output schemas in recursive CTEs and (3) writing the
CTE plans to the final plan output.  There may also be
some messing around with StmtCommonTableScans when joining
Derived Table scans, which we call From Subqueries, with
the larger queries.

Perhaps we can use this Common Table Expression mechanism
as a way to compute Derived Table Subqueries only when needed,
and only once.