Much wrestling with differentiators and column resolution.
We still replicate the display columns and the columns in the
table scan.  This seems to happen because we process them twice,
once for the base query and once for the recursive query.

These should both come from the output schema, so we need to
do something to subvert defining these when scanning the common
table XML.