Shoved around best cost plans.  Add plans to SeqScanPlanNode.
We need to be able to find plans when we create execution lists.
For subquery expressions we find them in the expressions.  For
common table queries we need to stick them someplace outside of
the planner machinery, since that's gone.  The plan tree is really
the only place for them, and the sequential scan nodes which scan
them is the logical place.  We need to remember which execution list
we use, though, and only add them once.