ENG-12820: (#5071)
* ENG-12820:
We don't currently assume the MP Memory context when creating Join Views because replicated tables joined with partitioned tables requires us to assume other contexts. However, the statements inside the view need to be allocated in the MP memory context so we assume the memory context after the handler has been registered with all the site sources. Since we assume contexts in a very localized way with views, we may want to consider a global assignment of the MP memory context after all the tests are passing and we understand the dynamics of all the required context switches.

There was also an issue with the deconstructor of the persistent table where JoinViews of replicated tables needed to be cleaned up in the partitioned Memory context. To support this scenario, a new scoped class was introduced that assumes the localSite context if we are in the MP Memory context.

* ENG-12820:
Fixed a few more places where indexes or tables are being modified in LoadCatalog, and the tables are being migrated rather than just deleted or inserted.