<voltdb>/tests/test-tools.sh: added a new script that consists of the generic parts of the old <voltdb>/tests/sqlgrammar/run.sh, which could apply to other test frameworks as well. It also incorporates (a modified version of) the wait-until-VoltDB-server-finishes-initializing parts of the old <voltdb>/tests/geb/vmc/server/run_voltdb_server.sh.
<voltdb>/tests/sqlgrammar/run.sh: modified to use the new
../test-tools.sh wherever possible; also improved the help option; and
other minor improvements.

<voltdb>/tests/geb/vmc/run.sh: added a new script that uses the new
../../test-tools.sh, and incorporates the still-relevant parts of both
server/run_voltdb_server.sh and run_vmc_tests.sh (which are now
deleted), allowing you to run the GEB tests of the VMC, including
building VoltDB, starting the server, etc., from a single script. Note:
the default browser changed from Firefox to Chrome; and the old options
of running against the Voter or Genqa test apps were eliminated.

Moved the deployment.xml and deploy_pro.xml config files to the
<voltdb>/tests/geb/vmc/ directory, from the ./server/ directory, which
is going away; and updated the latter to work with the current config
file format. Similarly, moved ddl.sql to the <voltdb>/tests/geb/vmc/
directory, from the ./server/ directory, and added some DROP statements
to the beginning, so that it can be run multiple times in a row.

fullDDL.sql: added this line:
DROP FUNCTION IntFunction IF EXISTS;
so that the related VMC tests can be run multiple times in a row.
(Tables and other things get dropped programmatically, but UDFs
(user-defined functions) are not displayed in the VMC.)

Updated the <voltdb>/tests/geb/vmc/README.md file.