ENG-10421: allow multiple hosts trying to rejoin at the same time (#3619) (#5089)
On startup, each node sends a request to rejoin the mesh to the live node. The live node will check if the request can be accepted at the moment. For example, we only allow one node to rejoin at a time. If more than one node tries to rejoin, the first rejoning node will atomically check and set the ZK blocker. Subsequent nodes will fail to do so and be rejected to join the mesh. They will retry after a while until they can succeed.

Change-Id: I2ccced38059fdd397b11872e5d3f228d52adc31e