ENG-12823: (#5096)
When multiple source hosts are involved in a rejoin, knowing when a sink has received all snapshot data is more complicated because the partitioned and replicated table data for one partition can come from 2 hosts. To resolve this all rejoin partitions except the lowest site will receive 2 EOF rejoin messages (One from the replicated table source and one from the partitioned table source). When both EOFs are received, the sink for the partition is cleaned up.

Also, in some cases the source of the replicated table for a rejoin may not be the lowest site of a host. This is problematic because it was previously assumed that only the lowest site could stream the replicated table data. Now any site can stream the replicated table data, but there is an extra check to verify that a new stream is not started until the old one has been completed. This check can be used to verify that 2 partitions don't attempt to stream the replicated table data at the same time.