sql-grammar.txt: added a new section defining grammar rules for Common Table Expressions (CTEs) - for now, just Recursive CTEs; and added CTE statements to the list of possible types of SQL statements; added a new VARCHAR column (see below for details), and simplified future adding of columns; added a small section defining JSON values (while I was messing with VARCHAR columns; needs more work, though); and a couple of tweaks of IPV4/6 values (VARCHAR and VARBINARY); improved some of the definitions related to joins (since they were messing up CTEs that use joins), particularly around using the same table, view, or alias name more than once in the same query; added 'rarely-null' versions of each data type (e.g. 'str-rarely-null-value'), for cases (like CTEs) where we want NULL values to be tested occasionally, but not normally used, since they are likely to produce invalid SQL statements; various other minor changes / improvements; and updated / improved some comments.
DDL.sql, JSP*.java, GetOrInsertBase.java: just added one column to each
table (and view): the old VCHAR column (16 chars or 64 BYTES) is renamed
VCHAR_OUTLINE_MIN; the new column has the old name of VCHAR, but is of
unspecified size (i.e., the default size) - which is what PostgreSQL
requires for RECURSIVE CTE queries.

sql_grammar_generator.py: increased the time to wait for sqlcmd to
"hang", from 10 to 15 seconds, so that it exceeds the standard query
timeout of 10 seconds; added the ability to escape curly braces ('{' and
'}') and backslashes ('\'), by adding a backslash in front of it (this
is in addition to the existing escaping of brackets, '[' and ']'); and
moved the echo / grammar symbols code outside of the if sqlcmd, so that
it works even when SQL statements are being generated without being
passed to sqlcmd.