Clear out all the local tables before doing HSQL operations.
Also some small fixes.  I removed the debug printing which Ethan
commented out, as it's not really useful anymore.  Also, when the list
of table names had a different size than the list of column types in the
base query it makes no sense.  But we didn't check and sometimes got an
ArrayIndexOutOfBoundsError or else a syntax error.