ENG-13317: Coordinate the processing of exceptions across site threads.
When an exception is thrown in the lowest site thread, the other threads detect the change and throw a generic exception as well. The generic exception is forwarded to the MPProcedureTask which ignores the exception in favor of the exception generated by the lowest site thread (on one of the hosts).