Fix ENG-13536, which is an orderby/union bug.
We tried to do an optimization using the display list of the left child
of a union statement.  Unfortunately we chose the orderby list of the
left child, not the display list.  This caused the orderby list of the
left child to be corrupted by changing its sort order.

This commit also beefs up the function for validating plans.