Eng 13536 Fix a problem with UNION and Sort Order (#5133)
We tried to do an optimization using the display list of the left child
of a union statement.  Unfortunately we chose the orderby list of the
left child, not the display list.  This caused the orderby list of the
left child to be corrupted by changing its sort order.

This commit also considerably beefs up the function for validating plans.