Fix problems with spaces in javascript.
Sometimes strings with spaces which have been scraped from
HTML DOM objects have characters which are not the ascii space
character but which look like spaces, or are not rendered at all.
If we compare these to ordinary ascii spaces we don't get what
we expect.  This causes apparent data corruption.  It's not really
data corruption, but it appears to be.  We look for this in
SQL string data and try to fix it.

This commit also compresses javascript files only in debug builds.
This makes debugging of javascript possible, though it's still not
easy.