Solve ENG-13484 just with .css files.
Apparently the browsers take some liberties with white space where they
are allowed to.  In the elements for SQL statements and in the elements
which display the results of SQL queries we really don't want them to do
that.  So we nail down the white space treatment to be "pre".  This
means no collapsing of white space, no replacing of spaces with non
breaking or non displaying space characters and no line or word breaks.

This may solve ENG-13003 as well.