ENG-13586: SQL-grammar-gen hanging on Recursive CTE infinite loop
sql_grammar_generator.py: when sqlcmd output includes 'Server is paused
and is available in read-only mode', end execution immediately with a
fatal error; moved (in the summary file and to STDOUT) the detailed
"hanging SQL" messages to before the last-n and summary messages, with
only a single line about "hanging SQL" after them (when such messages
appear); added 'Resource limit exceeded' to the list of ERROR messages
to be searched for, in the log and in the console output; in
get_last_n_sql_statements method, added optional times (that SQL was
sent and output received) to be printed out; and added code to save
these times, when we have "hanging SQL"; and therefore moved the
formatted_time method earlier, to be used in print_sql_statements.

sql-grammar.txt: added a "DEPTH" column to all Recursive CTE queries,
and always check that the DEPTH does not exceed a specified "max-depth"
(which is typically 5, but can range from 0 to 9), in order to avoid
infinite loops.

sql_grammar_generator.py: just corrected a comment, and improved an
error message.