Address Chris' review comments.
Make the siteId in the LargeTempTableBlockId be 64 bits.  This means we
can't have a single 64 bit integer for it.  But we need to pass it and
the block id up to Java through the JNI interface.  So we still need it
to be unsigned.

Change the Java LargeTempTableBlockManager to keep structured keys for
blocks rather than simple integers.  We could have done this with Java's
BigInteger and some arithmetic, but that seems cluttered.