Squash commits from Bill's work to fix small issues with CMake build, from pull request #5172
Make this work on C6.

Revert to the old valgrind parameters.  This gives worse
coverage, but it seems to work on C6.

Add a comment.

This is mostly to try to force a Jenkins run.

Add a comment.

This is mostly to try to force a Jenkins run.

Pass VOLT_LOG_LEVEL down properly.

Don't set it if it's already set.

Fix a typo in a variable name.