ENG-13677 C6 Memcheck failures (#5172)
Fix C6 Memcheck failures.

Memcheck was failing on C6 because of unsupported parameters to valgrind.  We now run Valgrind in the same way we used to run it.  This is not as useful as it could be, but it seems to be ok.

I also fixed setting VOLT_LOG_LEVEL from ant and fixed a variable name misspelling.

https://issues.voltdb.com/browse/ENG-13677