Eng 13658 gram cte nonrec (#5173)
* ENG-13658: sql-grammar.txt: added "basic" (non-recursive) CTE
statements; tweaked some Recursive CTE definitions to make them usable
for non-recursive CTEs too, mainly by using {depth}, {one-as-depth}, and
{aggregate-depth-funcs}, which can be used as before for Recursive CTEs,
but set to an empty string for non-recursive; and fixed a couple
comments.

sql_grammar_generator.py: improved the code involving symbol_reuse to
allow an empty string to be reused; and to make sure that a recursive
call to get_one_sql_statement does not ignore existing symbols; and to
include such values in "--echo" debug print.

* sql-grammar.txt: no material changes (with one exception): moved
definitions that are used by both Recursive and "basic" (non-recursive)
CTE statements to a section between the definitions for the "basic" and
Recursive CTEs; changed "rcte" to "cte" and "rec-cte-col" to "cte-col",
etc., since these are now used in both recursive and non-recursive CTEs
(but definitions used only by Recursive CTEs are unchanged); added some
new comments and updated and deleted some outdated ones. The one
exception is the definitions of cte-colX-alias (formerly
rec-cte-colX-alias), where X is 6, 7, 8, or 9, which are now correctly
defined to be equal to "[AS ]{cte-col-aliasX:ctecX}", whereas they used
to be defined as "[AS ]{rec-cte-col-aliasY:rctecX}", where Y = X - 5
(presumably due to a copy-paste error).